DISSERTATION: added rest of plots
plots added:
  emsa-selex
   - radial plots
   - kmer plots for enrichment of mononucleotides in each library
  pcr-bias
   - kmer plots for enz bias
   - kmer plot for purification

- also wrote something for biases from pcr cycles. i don't plan to
  include this in the main results as the results are dodgy, but it
  could be in the discussion. till i decide the text is in a big
  comment (it's definitely going to show in the lab meeting)

- renamed a picture so that it follows my naming scheme

it looks like i don't need to re-generate the kmer plots as the
resolution is high enough. however i think the fft-related plots
need to be replaced with the high-res ones from the server. that
can be done later though -- in the interests of having a 1st draft
this is not necessary yet