DISSERTATION: made numbers in table sci not; result figs done
at first i wanted to mess around with the settings in pgfplot to
make that happen. but then it misbehaved around the strings in the
csv file. then i discovered that manually making a new column is
much much much much easier

i also removed a column in the csv file to make this happen

a story can be told with the git log: i think it took more like 3
minutes to insert the table and... 1 hour to get the numbers to
display the way i want. this is latex for you

in other news i'm done with the figures in my results section,
spare the captions