DISSERTATION: edited second draft according to Fangjie's comments
Still not calling this a 3rd/final draft because I'm yet to ask
him about a sentence he added.  This sentence is about my description
of the k-mer plots that showed that bound and unbound libraries
suggested the same sequence preferences.  I'm not sure what he meant
by 'heavier PCR', and I think it might be misleading too.