DISSERATATION: clarified `problem sentence' for k-mer counts
Refer to last commit.

The unbound band on the EMSA gel is brighter than the bound band,
i.e. the PCR template for the unbound band is of a higher
concentration.  Therefore, even if they were amplified for the
same number of cycles, the unbound libraries were amplified less.
This is because the final PCR product is of a fixed concentration,
so the extent of amplification depends on the template concentration

Therefore this explains why there is more PCR bias for the bound
library.